Extract PasswordBoxBehavior into its own class
