+++ MaterialDesignThemes.Wpf/Themes/MaterialDesignTheme.PasswordBox.xaml
@@ -448,7 +448,7 @@
     <Setter Property="wpf:StylizedBehaviors.Behaviors">
       <Setter.Value>
         <wpf:StylizedBehaviorCollection>
-          <wpf:PasswordBoxAssist />
+          <internalbehaviors:PasswordBoxBehavior />
         </wpf:StylizedBehaviorCollection>
       </Setter.Value>
     </Setter>
@@ -958,7 +958,7 @@
     <Setter Property="wpf:StylizedBehaviors.Behaviors">
       <Setter.Value>
         <wpf:StylizedBehaviorCollection>
-          <wpf:PasswordBoxAssist />
+          <internalbehaviors:PasswordBoxBehavior />
         </wpf:StylizedBehaviorCollection>
       </Setter.Value>
     </Setter>
